basic Nm template engine
This engine renders the given path, in the source path, using Nm.
Sanford then takes the rendered template data and serializes it to
BSON as needed.

This engine memoizes the Nm source and binds to an overridable
scope name method, 'view'.

This should be all that is needed to have a working sanford Nm
template engine.