update to support locals
Sanford's engine api now takes locals.  This updates the engine to
honor locals that are passed to it.

This also updates the handling of the passed service handler.  This
is no longer referred to generically as "scope" but instead specifically
as "service handler".  This option to override the handler local
name has also been updated to better intend its purpose.

This just makes the engine compatible with the latest Sanford engine
API.