Pass sanford template engine ext to nm source
This updates the template engine to pass its ext to nm source. The
template engine ext is set by sanford when the engine is
registered. This makes it so nm will use templates with the
extension that was used when registering the engine with sanford.
This is the expected behavior and avoids confusion with nm randomly
not using a template that matches the registered extension.

This is part of fixing an issue with sanford not rendering the
expected nm template. This is because the nm engine no longer
forced templates extension to be `nm` (see redding/nm#16). By not
expecting an extension, the nm engine can't pick the correct file
as the template if there are multiple files with the same name
except for their extensions. For example, a common pattern is to
name a handler file and the template the same only the handler will
have `.rb` for its extension and the template would use `.nm`. In
this scenario, it's possible for the nm engine to try and use the
ruby file as its template. Nm is being updated to use a custom
extension (the one sanford-nm will now pass to it) to address this
issue (redding/nm#17).