implement `partial` engine method
Sanford has added this new API method to its engine.  It should
render a given template file given just locals.  This is needed
to stay compatible with the Sanford engine spec.

This updates the gemspec to force the latest Sanford, which adds
this new api method, and the latest Nm just to keep up with the
latest stuff.